Avocet journals module
Importing journals is only enabled for global- and tenant admins. To import a set of journals, a `POST` request needs to be sent to `/api/journals/import`. The request must contain a `.CSV`-file where the first 'column' contains the name of the journal (e.g. "Academic Research"). The second and third column contain the "print ISSN" and the "electronic ISSN".

Since we want to use the journals for an autosuggest feature in the front-end, we registered a search endpoint. Searching for journals can be done using the `/api/search/journals` endpoint. The endpoint accepts the 'q', 'limit', 'start' and 'sort' parameters (e.g. /api/search/journals?q=academic&limit=10).