Initial commit for tickets module
Tickets are automatically generated when creating a publication. The publications module can still work independently from the tickets module. Instead of using dispatched events or public functions to generate a ticket, we register a `createPublicationListener` on initialisation of the `oae-tickets` module. This listener gets executed when a publication was successfully created (see `_createPublication` in the `oae-publications` api).