Deleting and updating tickets
Tickets can be deleted by sending a request to
/api/ticket/:ticketId/delete. Only global admins and admins where the
ticket was created in can delete tickets.

There is currently no endpoint for updating tickets, since this is
handled by ZenDesk. When a ZenDesk ticket gets updated, this will be
synced with our tickets. Therefore, a public method has been created
(updateTicket) which accepts a ticketId and an opts object. The opts
object contains status and lastModified. All other parameters that are
passed in will be stripped from the opts object.